@@ -14,10 +14,10 @@ authors = [
   { name = "KeriBella", email = "keribella@git.com" },
   { name = "Daniel Balparda", email = "balparda@gmail.com" }
 ]
-requires-python = ">=3.11"
+requires-python = ">=3.13.5"  # if version changes, see .vscode/tasks.json & README.md
 dependencies = [
   # Baselib pulled straight from a tagged commit on GitHub
-  "baselib @ git+https://github.com/balparda/baselib.git@1.2",
+  "baselib @ git+https://github.com/balparda/baselib.git@1.3"
 ]
 
 # test with:
