@@ -1,4 +1,4 @@
-#!/usr/bin/python3 -O
+#!/usr/bin/env python3
 #
 # Copyright 2025 Balparda (balparda@github.com)
 # GNU General Public License v3 (http://www.gnu.org/licenses/gpl-3.0.txt)
@@ -38,8 +38,9 @@ __version__ = (1, 0)
 
 # defaults
 _DEFAULT_DAYS_FRESHNESS = 10
+_DAYS_CACHE_FRESHNESS = 1
 _SECONDS_IN_DAY = 60 * 60 * 24
-_DAYS_OLD: Callable[[float], float] = lambda t: (time.time() - t) / _SECONDS_IN_DAY
+DAYS_OLD: Callable[[float], float] = lambda t: (time.time() - t) / _SECONDS_IN_DAY
 DEFAULT_DATA_DIR: str = base.MODULE_PRIVATE_DIR(__file__, '.tfinta-data')
 _DB_FILE_NAME = 'transit.db'
 IRISH_RAIL_OPERATOR = 'Iarnród Éireann / Irish Rail'
@@ -50,7 +51,7 @@ _DATETIME_OBJ: Callable[[str], datetime.datetime] = lambda s: datetime.datetime.
     s, '%Y%m%d')
 # _UTC_DATE: Callable[[str], float] = lambda s: _DATETIME_OBJ(s).replace(
 #     tzinfo=datetime.timezone.utc).timestamp()
-_DATE_OBJ: Callable[[str], datetime.date] = lambda s: _DATETIME_OBJ(s).date()
+DATE_OBJ: Callable[[str], datetime.date] = lambda s: _DATETIME_OBJ(s).date()
 
 # type maps for efficiency and memory (so we don't build countless enum objects)
 _LOCATION_TYPE_MAP: dict[int, dm.LocationType] = {e.value: e for e in dm.LocationType}
@@ -89,6 +90,36 @@ class _TableLocation:
 # useful aliases
 _GTFSRowHandler = Callable[
     [_TableLocation, int, dict[str, Union[None, str, int, float, bool]]], None]
+
+
+def HMSToSeconds(time_str: str) -> int:
+  """Accepts 'H:MM:SS' or 'HH:MM:SS' and returns total seconds since 00:00:00.
+
+  Supports hours ≥ 0 with no upper bound. Very flexible, will even accept 'H:M:S' for example.
+
+  Args:
+    time_str: String to convert ('H:MM:SS' or 'HH:MM:SS')
+
+  Raises:
+    ValueError: malformed input
+  """
+  try:
+    h_str, m_str, s_str = time_str.split(':')
+  except ValueError as err:
+    raise ValueError(f'bad time literal {time_str!r}') from err
+  h, m, s = int(h_str), int(m_str), int(s_str)
+  if not (0 <= m < 60 and 0 <= s < 60):
+    raise ValueError(f'bad time literal {time_str!r}: minute and second must be 0-59')
+  return h * 3600 + m * 60 + s
+
+
+def SecondsToHMS(sec: int) -> str:
+  """Seconds from midnight to 'HH:MM:SS' representation. Supports any positive integer."""
+  if sec < 0:
+    raise ValueError(f'no negative time allowed, got {sec}')
+  h, sec = divmod(sec, 3600)
+  m, s = divmod(sec, 60)
+  return f'{h:02d}:{m:02d}:{s:02d}'
 
 
 class GTFS:
@@ -169,7 +200,7 @@ class GTFS:
       logging.info('Saved DB to %r (%s)', self._db_path, tm_save.readable)
 
   @functools.lru_cache(maxsize=1 << 14)
-  def _FindRoute(self, route_id: str) -> Optional[dm.Agency]:
+  def FindRoute(self, route_id: str) -> Optional[dm.Agency]:
     """Find route by finding its Agency."""
     for agency in self._db.agencies.values():
       if route_id in agency.routes:
@@ -177,13 +208,37 @@ class GTFS:
     return None
 
   @functools.lru_cache(maxsize=1 << 16)
-  def _FindTrip(self, trip_id: str) -> tuple[Optional[dm.Agency], Optional[dm.Route]]:
-    """Find route by finding its Agency & Route."""
+  def FindTrip(self, trip_id: str) -> tuple[
+      Optional[dm.Agency], Optional[dm.Route], Optional[dm.Trip]]:
+    """Find route by finding its Agency & Route. Return (agency, route, trip)."""
     for agency in self._db.agencies.values():
       for route in agency.routes.values():
         if trip_id in route.trips:
-          return (agency, route)
-    return (None, None)
+          return (agency, route, route.trips[trip_id])
+    return (None, None, None)
+
+  @functools.lru_cache(maxsize=1 << 10)
+  def StopName(self, stop_id: str) -> tuple[Optional[str], Optional[str], Optional[str]]:
+    """Gets (code, name, description) for a Stop object of `id`."""
+    if stop_id not in self._db.stops:
+      return (None, None, None)
+    stop: dm.BaseStop = self._db.stops[stop_id]
+    return (stop.code, stop.name, stop.description)
+
+  def ServicesForDay(self, day: datetime.date) -> set[int]:
+    """Return set[int] of services active (available/running/operating) on this day."""
+    weekday: int = day.weekday()
+    services: set[int] = set()
+    # go over available services
+    for service, calendar in self._db.calendar.items():
+      if calendar.start <= day <= calendar.end:
+        # day is in range for this service; check day of week and the exceptions
+        weekday_service: bool = calendar.week[weekday]
+        service_exception: Optional[bool] = calendar.exceptions.get(day)
+        has_service: bool = service_exception if service_exception is not None else weekday_service
+        if has_service:
+          services.add(service)
+    return services
 
   def FindAgencyRoute(
       self, agency_name: str, route_type: dm.RouteType, short_name: str,
@@ -218,10 +273,38 @@ class GTFS:
           return (agency, route)
     return (agency, None)
 
+  def LoadData(
+      self, operator: str, link: str,
+      freshness: int = _DEFAULT_DAYS_FRESHNESS, force_replace: bool = False) -> None:
+    """Downloads and parses GTFS data.
+
+    Args:
+      operator: Operator for GTFS file
+      link: URL for GTFS file
+      freshness: (default 10) Number of days before data is not fresh anymore and
+          has to be reloaded from source
+      force_replace: (default False) If True will parse a repeated version of the ZIP file
+    """
+    # first load the list of GTFS, if needed
+    if (age := DAYS_OLD(self._db.files.tm)) > freshness:
+      logging.info('Loading CSV sources (%0.2f days old)', age)
+      self._LoadCSVSources()
+    else:
+      logging.info('CSV sources are fresh (%0.2f days old) - SKIP', age)
+    # load GTFS data we are interested in
+    if (not force_replace and operator in self._db.files.files and
+        link in self._db.files.files[operator] and
+        (age := DAYS_OLD(self._db.files.files[operator][link].tm)) <= freshness):  # type:ignore
+      logging.info('GTFS sources are fresh (%0.2f days old) - SKIP', age)
+    else:
+      logging.info('Parsing GTFS ZIP source (%0.2f days old)', age)
+      self._LoadGTFSSource(operator, link, force_replace=force_replace)
+
   def _InvalidateCaches(self) -> None:
     """Clear all caches."""
-    self._FindRoute.cache_clear()
-    self._FindTrip.cache_clear()
+    self.FindRoute.cache_clear()
+    self.FindTrip.cache_clear()
+    self.StopName.cache_clear()
 
   def _LoadCSVSources(self) -> None:
     """Loads GTFS official sources from CSV."""
@@ -292,13 +375,33 @@ class GTFS:
     # load ZIP from URL
     done_files: set[str] = set()
     file_name: str
+    cache_file_name: str = link.replace('://', '__').replace('/', '_')
+    cache_file_path: str = os.path.join(self._dir_path, cache_file_name)
+    save_cache_file: bool
     with self._ParsingSession():
-      with urllib.request.urlopen(link) as gtfs_zip:
-        # extract files from ZIP
+      if (not force_replace and os.path.exists(cache_file_path) and
+          (age := DAYS_OLD(os.path.getmtime(cache_file_path))) <= _DAYS_CACHE_FRESHNESS):
+        # we will used the cached ZIP
+        logging.warning('Loading from %0.2f days old cache on disk! (use -r to override)', age)
+        url_opener = open(cache_file_path, 'rb')
+        save_cache_file = False
+      else:
+        # we will re-download from the URL
+        url_opener = urllib.request.urlopen(link)
+        save_cache_file = True
+      # open from whatever source
+      with url_opener as gtfs_zip:
+        # get ZIP binary content, and if we got from URL save to cache
         gtfs_zip_bytes: bytes = gtfs_zip.read()
         logging.info(
-            'Loading %r data, %s,from %r',
-            operator, base.HumanizedBytes(len(gtfs_zip_bytes)), link)
+            'Loading %r data, %s, from %r%s',
+            operator, base.HumanizedBytes(len(gtfs_zip_bytes)),
+            link if save_cache_file else cache_file_name,
+            ' => SAVING to cache' if save_cache_file else '')
+        if save_cache_file:
+          with open(cache_file_path, 'wb') as cache_file_obj:
+            cache_file_obj.write(gtfs_zip_bytes)
+        # extract files from ZIP
         for file_name, file_data in _UnzipFiles(io.BytesIO(gtfs_zip_bytes)):
           file_name = file_name.strip()
           location = _TableLocation(operator=operator, link=link, file_name=file_name)
@@ -395,9 +498,13 @@ class GTFS:
     self._changed = True
     logging.info('Read %d records from %s', i + 1, file_name)
 
+  ##################################################################################################
+  # GTFS ROW HANDLERS
+  ##################################################################################################
+
   # HANDLER TEMPLATE (copy and uncomment)
   # def _HandleTABLENAMERow(
-  #     self, location: _TableLocation, count: int, row: dict[str, Optional[str]]) -> None:
+  #     self, location: _TableLocation, count: int, row: dm.ExpectedFILENAMECSVRowType) -> None:
   #   """Handler: "FILENAME.txt" DESCRIPTION.
   #
   #   Args:
@@ -429,8 +536,8 @@ class GTFS:
       raise RowError(
           f'feed_info.txt table ({location}) is only supported to have 1 row (got {count}): {row}')
     # get data, check
-    start: datetime.date = _DATE_OBJ(row['feed_start_date'])
-    end: datetime.date = _DATE_OBJ(row['feed_end_date'])
+    start: datetime.date = DATE_OBJ(row['feed_start_date'])
+    end: datetime.date = DATE_OBJ(row['feed_end_date'])
     if start > end:
       raise RowError(f'incompatible start/end dates in {location}: {row}')
     # check against current version (and log)
@@ -500,8 +607,8 @@ class GTFS:
       RowError: error parsing this record
     """
     # get data, check
-    start: datetime.date = _DATE_OBJ(row['start_date'])
-    end: datetime.date = _DATE_OBJ(row['end_date'])
+    start: datetime.date = DATE_OBJ(row['start_date'])
+    end: datetime.date = DATE_OBJ(row['end_date'])
     if start > end:
       raise RowError(f'inconsistent row @{count} / {location}: {row}')
     # update
@@ -526,7 +633,7 @@ class GTFS:
     Raises:
       RowError: error parsing this record
     """
-    self._db.calendar[row['service_id']].exceptions[_DATE_OBJ(row['date'])] = (
+    self._db.calendar[row['service_id']].exceptions[DATE_OBJ(row['date'])] = (
         row['exception_type'] == '1')
 
   def _HandleRoutesRow(
@@ -593,7 +700,7 @@ class GTFS:
       RowError: error parsing this record
     """
     # check
-    agency: Optional[dm.Agency] = self._FindRoute(row['route_id'])
+    agency: Optional[dm.Agency] = self.FindRoute(row['route_id'])
     if agency is None:
       raise RowError(f'agency in row was not found @{count} / {location}: {row}')
     # update
@@ -664,8 +771,8 @@ class GTFS:
       raise RowError(f'invalid row @{count} / {location}: {row}')
     if row['stop_id'] not in self._db.stops:
       raise RowError(f'stop_id in row was not found @{count} / {location}: {row}')
-    agency, route = self._FindTrip(row['trip_id'])
-    if not agency or not route:
+    agency, route, trip = self.FindTrip(row['trip_id'])
+    if not agency or not route or not trip:
       raise RowError(f'trip_id in row was not found @{count} / {location}: {row}')
     # update
     self._db.agencies[agency.id].routes[route.id].trips[row['trip_id']].stops[row['stop_sequence']] = dm.Stop(
@@ -674,26 +781,35 @@ class GTFS:
         timepoint=row['timepoint'], headsign=row['stop_headsign'],
         pickup=pickup, dropoff=dropoff)
 
-  def LoadData(
-      self, operator: str, link: str,
-      freshness: int = _DEFAULT_DAYS_FRESHNESS, force_replace: bool = False) -> None:
-    """Downloads and parses GTFS data.
+  ##################################################################################################
+  # GTFS PRETTY PRINTS
+  ##################################################################################################
 
-    Args:
-      operator: Operator for GTFS file
-      link: URL for GTFS file
-      freshness: (default 1) Number of days before data is not fresh anymore and
-          has to be reloaded from source
-      force_replace: (default False) If True will parse a repeated version of the ZIP file
-    """
-    # first load the list of GTFS, if needed
-    if (age := _DAYS_OLD(self._db.files.tm)) > freshness:
-      logging.info('Loading CSV sources (%0.1f days old)', age)
-      self._LoadCSVSources()
-    else:
-      logging.info('Sources are fresh (%0.1f days old) - SKIP', age)
-    # load GTFS data we are interested in
-    self._LoadGTFSSource(operator, link, force_replace=force_replace)
+  def PrettyPrintTrip(self, trip_id: str) -> Generator[str, None, None]:
+    """Generate a pretty version of a Trip."""
+    agency, route, trip = self.FindTrip(trip_id)
+    if not agency or not route or not trip:
+      raise ValueError(f'trip id {trip_id!r} was not found')
+    yield f'ID:     {trip.id}'
+    yield f'Agency: {agency.name}'
+    yield f'Route:  {route.id}'
+    yield f'        Short name:  {route.short_name}'
+    yield f'        Long name:   {route.long_name}'
+    yield f'        Description: {route.description if route.description else "-"}'
+    yield f'Headsign:  {trip.headsign}'
+    yield f'Name:      {trip.name}'
+    yield f'Direction: {"inbound" if trip.direction else "outbound"}'
+    yield f'Service:   {trip.service}'
+    yield f'Shape:     {trip.shape}'
+    yield f'Block:     {trip.block}'
+    yield ''
+    yield '#    ARRIVAL  DEPART.  CODE        NAME'
+    for seq in sorted(trip.stops.keys()):
+      stop: dm.Stop = trip.stops[seq]
+      stop_code, stop_name, stop_description = self.StopName(stop.stop)
+      yield (f'{seq:03}: {SecondsToHMS(stop.arrival)} {SecondsToHMS(stop.departure)} '
+             f'@{stop.stop} {stop_code}/{stop_name}/'
+             f'{stop_description if stop_description else "-"}')
 
 
 def _UnzipFiles(in_file: IO[bytes]) -> Generator[tuple[str, bytes], None, None]:
@@ -719,36 +835,6 @@ def _UnzipFiles(in_file: IO[bytes]) -> Generator[tuple[str, bytes], None, None]:
         yield (file_name, file_data.read())
 
 
-def HMSToSeconds(time_str: str) -> int:
-  """Accepts 'H:MM:SS' or 'HH:MM:SS' and returns total seconds since 00:00:00.
-
-  Supports hours ≥ 0 with no upper bound. Very flexible, will even accept 'H:M:S' for example.
-
-  Args:
-    time_str: String to convert ('H:MM:SS' or 'HH:MM:SS')
-
-  Raises:
-    ValueError: malformed input
-  """
-  try:
-    h_str, m_str, s_str = time_str.split(':')
-  except ValueError as err:
-    raise ValueError(f'bad time literal {time_str!r}') from err
-  h, m, s = int(h_str), int(m_str), int(s_str)
-  if not (0 <= m < 60 and 0 <= s < 60):
-    raise ValueError(f'bad time literal {time_str!r}: minute and second must be 0-59')
-  return h * 3600 + m * 60 + s
-
-
-def SecondsToHMS(sec: int) -> str:
-  """Seconds from midnight to 'HH:MM:SS' representation. Supports any positive integer."""
-  if sec < 0:
-    raise ValueError(f'no negative time allowed, got {sec}')
-  h, sec = divmod(sec, 3600)
-  m, s = divmod(sec, 60)
-  return f'{h:02d}:{m:02d}:{s:02d}'
-
-
 def Main() -> None:
   """Main entry point."""
   # parse the input arguments, add subparser for `command`
@@ -764,14 +850,19 @@ def Main() -> None:
       '-r', '--replace', type=int, default=0,
       help='0 == does not load the same version again ; 1 == forces replace version (default: 0)')
   # "print" command
-  _: argparse.ArgumentParser = command_arg_subparsers.add_parser(
+  print_parser: argparse.ArgumentParser = command_arg_subparsers.add_parser(
       'print', help='Print DB')
+  print_arg_subparsers = print_parser.add_subparsers(dest='print_command')
+  trip_parser: argparse.ArgumentParser = print_arg_subparsers.add_parser(
+      'trip', help='Print Trip')
+  trip_parser.add_argument('-i', '--id', type=str, default='', help='Trip ID (default: "")')
   # ALL commands
   # parser.add_argument(
   #     '-r', '--readonly', type=bool, default=False,
   #     help='If "True" will not save database (default: False)')
   args: argparse.Namespace = parser.parse_args()
   command = args.command.lower().strip() if args.command else ''
+  print_command = args.print_command.lower().strip() if args.print_command else ''
   # start
   print(f'{base.TERM_BLUE}{base.TERM_BOLD}***********************************************')
   print(f'**                 {base.TERM_LIGHT_RED}GTFS DB{base.TERM_BLUE}                   **')
@@ -784,17 +875,22 @@ def Main() -> None:
     # execute the command
     print()
     with base.Timer() as op_timer:
-      # "read" command
-      if command == 'read':
-        database.LoadData(
-            IRISH_RAIL_OPERATOR, IRISH_RAIL_LINK,
-            freshness=args.freshness, force_replace=bool(args.replace))
-      # "print" command
-      elif command == 'print':
-        raise NotImplementedError()
-      # no valid command
-      else:
-        parser.print_help()
+      # look at main command
+      match command:
+        case 'read':
+          database.LoadData(
+              IRISH_RAIL_OPERATOR, IRISH_RAIL_LINK,
+              freshness=args.freshness, force_replace=bool(args.replace))
+        case 'print':
+          # look at sub-command for print
+          match print_command:
+            case 'trip':
+              for line in database.PrettyPrintTrip(args.id):
+                print(line)
+            case _:
+              raise NotImplementedError()
+        case _:
+          raise NotImplementedError()
       print()
       print()
     print(f'Executed in {base.TERM_GREEN}{op_timer.readable}{base.TERM_END}')
