@@ -1,4 +1,4 @@
-#!/usr/bin/python3 -bb
+#!/usr/bin/env python3
 #
 # Copyright 2025 Daniel Balparda (balparda@gmail.com)
 # GNU General Public License v3 (http://www.gnu.org/licenses/gpl-3.0.txt)
@@ -35,6 +35,6 @@ def Main() -> None:
   logging.info('*' * 80)
 
 
-if __name__ == "__main__":
+if __name__ == '__main__':
   logging.basicConfig(level=logging.INFO, format=base.LOG_FORMAT)
   Main()  # type:ignore
